requirements.txt
# Core Flask dependencies
blinker==1.9.0
click==8.1.8
colorama==0.4.6 ; platform_system == "Windows"
flask==3.1.1
flask-cors==4.0.0
importlib-metadata==8.7.0 ; python_version == "3.9"
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
werkzeug==3.1.3
zipp==3.23.0 ; python_version == "3.9"
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0
gevent==24.2.1
greenlet==3.0.3
zope.event==5.0
zope.interface==6.1